+++ core/src/main/scala/ai/h2o/sparkling/frame/H2OColumn.scala
@@ -1,0 +1,33 @@
+/*
+* Licensed to the Apache Software Foundation (ASF) under one or more
+* contributor license agreements.  See the NOTICE file distributed with
+* this work for additional information regarding copyright ownership.
+* The ASF licenses this file to You under the Apache License, Version 2.0
+* (the "License"); you may not use this file except in compliance with
+* the License.  You may obtain a copy of the License at
+*
+*    http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+package ai.h2o.sparkling.frame
+
+case class H2OColumn(
+                      name: String,
+                      dataType: H2OColumnType.Value,
+                      min: Double,
+                      max: Double,
+                      mean: Double,
+                      sigma: Double,
+                      numberOfZeros: Long,
+                      numberOfMissingElements: Long,
+                      percentiles: Array[Double],
+                      domain: Array[String],
+                      domainCardinality: Long) {
+  def nullable: Boolean = numberOfMissingElements > 0
+}
