+++ core/src/main/scala/ai/h2o/sparkling/frame/H2OFrame.scala
@@ -17,7 +17,13 @@
 
 package ai.h2o.sparkling.frame
 
-import org.apache.spark.h2o.utils.NodeDesc
+import java.text.MessageFormat
+import java.util
+
+import org.apache.spark.h2o.H2OContext
+import org.apache.spark.h2o.backends.external.RestApiUtils
+import org.apache.spark.h2o.backends.external.RestApiUtils.{getClusterEndpoint, getFrame, update}
+import water.api.schemas3.{RapidsFrameV3, SplitFrameV3}
 
 /* This case class contains metadata describing H2O Frame accessed via REST API */
 case class H2OFrame(
@@ -27,30 +33,45 @@ case class H2OFrame(
   lazy val numberOfRows: Long = chunks.foldLeft(0L)((acc, chunk) => acc + chunk.numberOfRows)
 
   def numberOfColumns: Int = columns.length
-}
 
-case class H2OColumn(
-                      name: String,
-                      dataType: H2OColumnType.Value,
-                      min: Double,
-                      max: Double,
-                      mean: Double,
-                      sigma: Double,
-                      numberOfZeros: Long,
-                      numberOfMissingElements: Long,
-                      percentiles: Array[Double],
-                      domain: Array[String],
-                      domainCardinality: Long) {
-  def nullable: Boolean = numberOfMissingElements > 0
-}
+  def convertAllStringColumnsToCategorical(): H2OFrame = {
+    val columns = this.columns.filter(_.dataType == H2OColumnType.string).map(_.name)
+    convertColumnsToCategorical(columns)
+  }
 
-object H2OColumnType extends Enumeration {
-  val enum, string, int, real, time, uuid = Value
+  def convertColumnsToCategorical(columns: Array[String]): H2OFrame = {
+    val conf = H2OContext.ensure().getConf
+    val endpoint = getClusterEndpoint(conf)
+    val indices = this.columns.map(_.name).zipWithIndex.toMap
+    val selectedIndices = columns.map { name =>
+      indices.getOrElse(name, throw new IllegalArgumentException(s"Column $name does not exist in the frame $frameId"))
+    }
+    val params = Map(
+      "ast" -> MessageFormat.format(s"( assign {0} (:= {0} (as.factor (cols {0} {1})) {1} []))", frameId, util.Arrays.toString(selectedIndices))
+    )
+    val rapidsFrameV3 = update[RapidsFrameV3](endpoint, "99/Rapids", conf, params)
+    getFrame(conf, rapidsFrameV3.key.name)
+  }
 
-  def fromString(dataType: String): Value = {
-    values.find(_.toString == dataType).getOrElse(throw new RuntimeException(s"Unknown H2O's Data type $dataType"))
+  def splitToTrainAndValidationFrames(splitRatio: Double): Array[H2OFrame] = {
+    val conf = H2OContext.ensure().getConf
+    if (splitRatio >= 1.0) {
+      throw new IllegalArgumentException("Split ratio must be lower than 1.0")
+    }
+    val endpoint = getClusterEndpoint(conf)
+    val params = Map(
+      "ratios" -> Array(splitRatio),
+      "dataset" -> frameId
+    )
+    val splitFrameV3 = update[SplitFrameV3](endpoint, "3/SplitFrame", conf, params)
+    splitFrameV3.destination_frames.map(frameKey => getFrame(conf, frameKey.name))
   }
 }
 
+object H2OFrame {
 
-case class H2OChunk(index: Int, numberOfRows: Long, location: NodeDesc)
+  def apply(frameId: String): H2OFrame = {
+    val conf = H2OContext.ensure().getConf
+    RestApiUtils.getFrame(conf, frameId)
+  }
+}
