+++ core/src/test/scala/ai/h2o/sparkling/FrameRestApiTestSuite.scala
@@ -16,9 +16,8 @@
 */
 package ai.h2o.sparkling
 
-import ai.h2o.sparkling.frame.{H2OColumnType, H2OFrameUtils}
+import ai.h2o.sparkling.frame.{H2OColumnType, H2OFrame}
 import org.apache.spark.SparkContext
-import org.apache.spark.h2o.backends.external.RestApiUtils
 import org.apache.spark.h2o.utils.SharedH2OTestContext
 import org.junit.runner.RunWith
 import org.scalatest.FunSuite
@@ -32,19 +31,18 @@ class FrameRestApiTestSuite extends FunSuite with SharedH2OTestContext {
       .set("spark.ext.h2o.rest.api.based.client", "true"))
 
 
-  private def uploadH2OFrame(): String = {
+  private def uploadH2OFrame(): H2OFrame = {
     // since we did not ask Spark to infer schema, all columns have been parsed as Strings
     val df = spark.read.option("header", "true").csv(TestUtils.locate("smalldata/prostate/prostate.csv"))
-    hc.asH2OFrameKeyString(df)
+    H2OFrame(hc.asH2OFrameKeyString(df))
   }
 
-  test("convertAllStringVecToCategorical") {
-    val h2oFrameKey = uploadH2OFrame()
-    val originalFrame = RestApiUtils.getFrame(hc.getConf, h2oFrameKey)
+  test("convertAllStringColumnsToCategorical") {
+    val originalFrame = uploadH2OFrame()
 
     val stringColumns = originalFrame.columns.filter(_.dataType == H2OColumnType.string).map(_.name)
     assert(stringColumns.nonEmpty)
-    val alteredFrame = H2OFrameUtils.convertAllStringVecToCategorical(h2oFrameKey)
+    val alteredFrame = originalFrame.convertAllStringColumnsToCategorical()
     val convertedColumns = alteredFrame.columns.filter(col => stringColumns.contains(col.name))
 
     // Verify that columns we asked to convert to categorical has been converted
@@ -58,10 +56,9 @@ class FrameRestApiTestSuite extends FunSuite with SharedH2OTestContext {
   }
 
   test("convertColumnsToCategorical") {
-    val h2oFrameKey = uploadH2OFrame()
-    val originalFrame = RestApiUtils.getFrame(hc.getConf, h2oFrameKey)
+    val originalFrame = uploadH2OFrame()
     val columnsToConvert = Array("ID", "AGE")
-    val alteredFrame = H2OFrameUtils.convertColumnsToCategorical(h2oFrameKey, columnsToConvert)
+    val alteredFrame = originalFrame.convertColumnsToCategorical(columnsToConvert)
     val convertedColumns = alteredFrame.columns.filter(col => columnsToConvert.contains(col.name))
     // Verify that columns we asked to convert to categorical has been converted
     convertedColumns.foreach { col =>
@@ -74,16 +71,16 @@ class FrameRestApiTestSuite extends FunSuite with SharedH2OTestContext {
   }
 
   test("splitFrameToTrainAndValidationFrames with ratio 1.0") {
-    val h2oFrameKey = uploadH2OFrame()
+    val originalFrame = uploadH2OFrame()
     val thrown = intercept[IllegalArgumentException] {
-      H2OFrameUtils.splitFrameToTrainAndValidationFrames(h2oFrameKey, 1.0)
+      originalFrame.splitToTrainAndValidationFrames(1.0)
     }
     assert(thrown.getMessage == "Split ratio must be lower than 1.0")
   }
 
   test("splitFrameToTrainAndValidationFrames with ratio lower than 1.0") {
-    val h2oFrameKey = uploadH2OFrame()
-    val splitFrames = H2OFrameUtils.splitFrameToTrainAndValidationFrames(h2oFrameKey, 0.9)
+    val originalFrame = uploadH2OFrame()
+    val splitFrames = originalFrame.splitToTrainAndValidationFrames(0.9)
     assert(splitFrames.length == 2)
     val train = splitFrames(0)
     val valid = splitFrames(1)
